+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Category;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
-use Intervention\Image\Image;
+use Image;
 use App\Http\Controllers\Controller;
 
 class CategoryController extends Controller
@@ -26,7 +26,7 @@ class CategoryController extends Controller
 
         $data = new Category();
         $data->name = $request->input('name');
-        $data->slug=Str::slug($request->input('name'));
+        $data->slug = Str::slug($request->input('name'));
 
         if ($request->hasFile('image')) {
             $file = $request->file('image');
@@ -55,10 +55,11 @@ class CategoryController extends Controller
     {
         $data = Category::findOrFail($id);
         $data->name = $request->input('name');
-        $data->slug=Str::slug($request->input('name'));
+        $data->slug = Str::slug($request->input('name'));
 
         if ($request->hasFile('image')) {
-            if ($data->image) {
+           
+            if ($data->image && file_exists(public_path('uploads/category') . '/' . $data->image)) {
                 unlink(public_path('uploads/category') . '/' . $data->image);
             }
 
@@ -68,7 +69,7 @@ class CategoryController extends Controller
             Image::make(public_path('uploads/category') . '/' . $filename)->resize(420, 260)->save('uploads/category/' . $filename);
             $data->image = $filename;
         }
-        
+
         $data->save();
 
         $notification = array(
@@ -83,7 +84,7 @@ class CategoryController extends Controller
     {
         $data = Category::findOrFail($id);
 
-        if ($data->image) {
+        if ($data->image && file_exists(public_path('uploads/category') . '/' . $data->image)) {
             unlink(public_path('uploads/category') . '/' . $data->image);
         }
